requirements.txt
moviepy==1.0.3
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.8.0
elevenlabs==1.5.0
groq==0.11.0
cohere==5.9.1
tqdm==4.66.1
g4f==0.3.3.0
openai==1.45.0
